fix linux CI

+++ src/utest/gc.cpp
@@ -32,7 +32,7 @@ static void string_test(aactor_t* a)
 {
     char buff[64];
     for (aint_t i = 0; i < 1000; ++i) {
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         any_push_string(a, buff);
     }
     for (aint_t i = 0; i < 1000; ++i) {
@@ -41,7 +41,7 @@ static void string_test(aactor_t* a)
         }
     }
     for (aint_t i = 1000; i < 5000; ++i) {
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         any_push_string(a, buff);
     }
     for (aint_t i = 1000; i < 5000; ++i) {
@@ -50,7 +50,7 @@ static void string_test(aactor_t* a)
         }
     }
     for (aint_t i = 5000; i < 10000; ++i) {
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         any_push_string(a, buff);
     }
     for (aint_t i = 5000; i < 10000; ++i) {
@@ -60,7 +60,7 @@ static void string_test(aactor_t* a)
     }
     for (aint_t i = 0; i < 10000; ++i) {
         if (i % 2 == 0) continue;
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         CHECK_THAT(any_to_string(a, i), Catch::Equals(buff));
     }
     any_push_string(a, "ok");
