+++ src/utest/msbox.cpp
@@ -149,11 +149,11 @@ static void string_producer_actor(aactor_t* a)
     char buff[32];
     for (aint_t i = 0; true; i += 2) {
         any_push_idx(a, -1);
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         any_push_string(a, buff);
         any_mbox_send(a);
         any_push_idx(a, -1);
-        snprintf(buff, sizeof(buff), "string %d", i + 1);
+        snprintf(buff, sizeof(buff), "string %d", (int)(i + 1));
         any_push_string(a, buff);
         any_mbox_send(a);
         any_yield(a);
@@ -169,7 +169,7 @@ static void string_consumer_actor(aactor_t* a)
         any_mbox_remove(a);
         REQUIRE(any_count(a) == 1);
         REQUIRE(any_type(a, 0).type == AVT_STRING);
-        snprintf(buff, sizeof(buff), "string %d", i);
+        snprintf(buff, sizeof(buff), "string %d", (int)i);
         CHECK_THAT(any_to_string(a, 0), Catch::Equals(buff));
     }
     done = true;
